Refactor automount functionality into separate package
Since we're adding more features to automounting, it makes sense to move
functionality for getting configmaps/secrets/pvcs to a separate package
from workspace provisioning.

This commit splits the only large automount file into files for common
functionality, managing secrets, and managing configmaps
